New Iterator Features

+++ examples/file_buffer.rs
@@ -14,7 +14,7 @@ fn main() {
     let chunk_size = 1024;
 
     let mut pb = Bar {
-        total: std::fs::metadata(src_file_path).unwrap().len() as usize,
+        total: std::fs::metadata(src_file_path).unwrap().len(),
         unit_scale: true,
         unit_divisor: 1024,
         unit: "B".to_string(),
@@ -27,7 +27,7 @@ fn main() {
         reader_ref.take(chunk_size).read_to_end(&mut chunk).unwrap();
         let chunk_len = chunk.len();
         dst_file.write(&chunk).unwrap();
-        pb.update(chunk_len);
+        pb.update(chunk_len as u64);
 
         if chunk_len == 0 {
             pb.refresh();
